services/vessel-observability/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database drivers
psycopg2-binary==2.9.9
cassandra-driver==3.28.0

# Kafka
kafka-python==2.0.2

# OpenTelemetry
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0

# Prometheus
prometheus-client==0.19.0

# Scientific computing for anomaly detection
numpy==1.26.2
scipy==1.11.4

# HTTP client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0